+++ crd-generator/test/src/test/java/io/fabric8/crd/generator/approvaltests/jackson3/Jackson3.java
@@ -15,12 +15,13 @@
  */
 package io.fabric8.crd.generator.approvaltests.jackson3;
 
+import io.fabric8.kubernetes.api.model.Namespaced;
 import io.fabric8.kubernetes.client.CustomResource;
 import io.fabric8.kubernetes.model.annotation.Group;
 import io.fabric8.kubernetes.model.annotation.Version;
 
 @Group("samples.fabric8.io")
 @Version("v1")
-public class Jackson3 extends CustomResource<Jackson3Spec, Void> {
+public class Jackson3 extends CustomResource<Jackson3Spec, Void> implements Namespaced {
 
 }