+++ extensions/istio/model/src/generated-builders/java/io/fabric8/istio/api/api/networking/v1alpha3/TrafficPolicyRetryBudgetBuilder.java
@@ -25,7 +25,7 @@ public class TrafficPolicyRetryBudgetBuilder extends TrafficPolicyRetryBudgetFlu
   }
 
   public TrafficPolicyRetryBudget build() {
-    TrafficPolicyRetryBudget buildable = new TrafficPolicyRetryBudget(fluent.getMinRetryConcurrency(), fluent.getPercent());
+    TrafficPolicyRetryBudget buildable = new TrafficPolicyRetryBudget(fluent.getBudgetInterval(), fluent.getMinRetryConcurrency(), fluent.getPercent());
     buildable.setAdditionalProperties(fluent.getAdditionalProperties());
     return buildable;
   }