+++ kubernetes-model-generator/openshift-model-monitoring/src/generated-builders/java/io/fabric8/openshift/api/model/monitoring/v1beta1/WebhookConfigFluent.java
@@ -22,6 +22,7 @@ public class WebhookConfigFluent<A extends io.fabric8.openshift.api.model.monito
   private Map<String,Object> additionalProperties;
   private HTTPConfigBuilder httpConfig;
   private Integer maxAlerts;
+  private String payload;
   private Boolean sendResolved;
   private String timeout;
   private String url;
@@ -67,6 +68,7 @@ public class WebhookConfigFluent<A extends io.fabric8.openshift.api.model.monito
     if (instance != null) {
         this.withHttpConfig(instance.getHttpConfig());
         this.withMaxAlerts(instance.getMaxAlerts());
+        this.withPayload(instance.getPayload());
         this.withSendResolved(instance.getSendResolved());
         this.withTimeout(instance.getTimeout());
         this.withUrl(instance.getUrl());
@@ -116,6 +118,9 @@ public class WebhookConfigFluent<A extends io.fabric8.openshift.api.model.monito
     if (!(Objects.equals(maxAlerts, that.maxAlerts))) {
       return false;
     }
+    if (!(Objects.equals(payload, that.payload))) {
+      return false;
+    }
     if (!(Objects.equals(sendResolved, that.sendResolved))) {
       return false;
     }
@@ -142,6 +147,10 @@ public class WebhookConfigFluent<A extends io.fabric8.openshift.api.model.monito
     return this.maxAlerts;
   }
   
+  public String getPayload() {
+    return this.payload;
+  }
+  
   public Boolean getSendResolved() {
     return this.sendResolved;
   }
@@ -166,6 +175,10 @@ public class WebhookConfigFluent<A extends io.fabric8.openshift.api.model.monito
     return this.maxAlerts != null;
   }
   
+  public boolean hasPayload() {
+    return this.payload != null;
+  }
+  
   public boolean hasSendResolved() {
     return this.sendResolved != null;
   }
@@ -183,7 +196,7 @@ public class WebhookConfigFluent<A extends io.fabric8.openshift.api.model.monito
   }
   
   public int hashCode() {
-    return Objects.hash(httpConfig, maxAlerts, sendResolved, timeout, url, urlSecret, additionalProperties);
+    return Objects.hash(httpConfig, maxAlerts, payload, sendResolved, timeout, url, urlSecret, additionalProperties);
   }
   
   public A removeFromAdditionalProperties(String key) {
@@ -221,6 +234,11 @@ public class WebhookConfigFluent<A extends io.fabric8.openshift.api.model.monito
     if (!(maxAlerts == null)) {
         sb.append("maxAlerts:");
         sb.append(maxAlerts);
+        sb.append(",");
+    }
+    if (!(payload == null)) {
+        sb.append("payload:");
+        sb.append(payload);
         sb.append(",");
     }
     if (!(sendResolved == null)) {
@@ -295,6 +313,11 @@ public class WebhookConfigFluent<A extends io.fabric8.openshift.api.model.monito
   
   public UrlSecretNested<A> withNewUrlSecretLike(SecretKeySelector item) {
     return new UrlSecretNested(item);
+  }
+  
+  public A withPayload(String payload) {
+    this.payload = payload;
+    return (A) this;
   }
   
   public A withSendResolved() {